Move RootStackParamList to Navigation/types
