+++ src/hooks/useFlipper/types.ts
@@ -1,5 +1,5 @@
 import {NavigationContainerRefWithCurrent} from '@react-navigation/core';
-import {RootStackParamList} from '@src/types/modules/react-navigation';
+import {RootStackParamList} from '@libs/Navigation/types';
 
 type UseFlipper = (ref: NavigationContainerRefWithCurrent<RootStackParamList>) => void;
